nasa-precip-engine/pyproject.toml
[project]
name = "nasa-precip-engine"
version = "1.0.0"
description = "NASA GPM IMERG precipitation microservice for GeoLens"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "earthaccess>=0.8.0",
    "xarray>=2024.1.0",
    "numpy>=1.26.0",
    "h3>=3.7.0",
    "pydantic>=2.5.0",
    "netCDF4>=1.6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
